+++ component/frontend/Controller/Subscribe.php
@@ -118,9 +118,11 @@ class Subscribe extends Controller
 		if (!$result)
 		{
 			$this->container->platform->setSessionVar('firstrun', false, 'com_akeebasubs');
+			$helpCode = basename($model->getLogFilename(), '.php');
 
 			$url = str_replace('&amp;', '&', \JRoute::_('index.php?option=com_akeebasubs&view=Level&layout=default&slug=' . $model->slug));
 			$msg = \JText::_('COM_AKEEBASUBS_LEVEL_ERR_VALIDATIONOVERALL');
+			$msg = \JText::sprintf('COM_AKEEBASUBS_LEVEL_ERR_VALIDATIONOVERALL_HELPCODE', $helpCode);
 
 			$this->setRedirect($url, $msg, 'error');